AGRICULTURAL-BIOLOGY(SEL'SKOKHOZYAISTVENNAYA-BILOGIYA).csl
The AGRICULTURAL BIOLOGY (having been printed since 1935 till 1966 under the title of Agrobiologiya [Agrobiology]) was established as a scientific periodical. It is the only academic journal in Russia which is completely focused on the fundamental research in classical and modern biology of agricultural plants, animals, microorganisms, their improvement and effective use with minimal adverse impact on the environment. The journal is published bimonthly, with six issues a year.